Support dictionary encoding for strings (#3681)
* initial dict encoded array data type

* unbox Arrow dict array for testing

* boxing for dict array

* basic array attrs/methods

* underscore in wrapper names

* box into numpy array of string objects

* flag to unbox to dict-encoded array for testing

* Support nbytes for dict str arr

* Support constant lowering for dict array

* bool arr indexing for dict array

* int arr indexing for dict arr

* simplify indexing

* bin ops for dict str arr

* add for dict str arr

* str int mul for dict array

* basic operations for dict arr

* add iterator for dict arr

* pd.to_datetime for dict arr

* fix notna() for array types

* handle dict str arr in Index, fix index/binary arr iterators

* fallback to regular str for shift()

* pd.to_numeric for dict arr

* Series.str.len() for dict arr

* Series.str.get/join for dict arr

* remove unused function

* delete dead code from previous join impl

* setna for dict arr

* concat() for dict arr

* Series.str.split() fix for dict array

* handle dict arr in alloc_type and gatherv

* scatterv for dict arr

* array to info for dictionary encoded string array

* refactor info_to_array codegen

* info_to_array for dict arr

* decref_array/incref_array for dict arr

* add has_global_dictionary flag to dict array

* clang-format on cpp changes

* np.repeat for dict arr

* has_global_dictionary in const lowering

* fix Series.fillna()

* avoid dict arr unboxing for nested str arrs

* handle index_from_array, remove some dead code

* dict arr in conversion functions

* dict arr in series/dist pass

* int getitem bcast fix

* remove dead code

* error for inplace fillna

* fix Series.replace()

* support dict arr mix in pd.concat()

* avoid dict arr conversion in non-unboxing context

* fix Series argsort

* fallback for df.sort_values

* raise clear error for inplace sort of dict arr

* handle decode_if_dict_array in opt, dist analysis

* handle dict arr in isin

* fallback in groupby for dict arr

* groupby type fixes

* fix groupby apply

* fix groupby udf dict arr type

* fix df.dtypes test

* fixes for read IO

* where/mask dict arr

* decode bug fix

* fix Series.str methods

* Series.dropna() fix for dict arr

* fix str search regex

* error on dict arr mutation

* fix table gatherv for dict arr

* fix confusion_matrix

* fix label encoder

* handle dict arr in init string index

* bugfix in read_excel

* dictionary-encoded array handling in C++ (#3688)

* dictionary-encoded array handling in C++

* fixes

* some improvements

* code formatting

* optimize array filter for dict arr

* optimize A.astype("Int64") for dict arr

* Skip tests that don't support dict arr

* avoid memory leak for dict arr in array_to_info

* [BE-2343] `read_parquet` implementation for DictionaryArray (#3687)

* monkey-patch _dataset_parquet to access reader_props

* Remove dead code

* skeleton code for reading in dict encoded arrays

* Fix typo

* Manual formatting on pyx file

* Unify all chunks

* Add pseudo code for dict-arr col determination

* get read_parquet IO working and fill gaps

* cleanup

* minor


* drop_duplicates and join gaps and other improvements

* fix gatherv bug for over-allocated string arrays

* skip fillna inplace test

* some fixes in C++

* another null check gap in C++ for dict

* fix binary index bug

* skip unsupported tests like setitem/fillna inplace

* Apply suggestions from code review


* code review suggestions

* Heuristic to decide when to read string columns as dictionary array (#3699)

* TestEqualColumn for DICT array

* proper return for alloc_dict_string_array

* Added initial changes for using dict array in a join cfunc (#3701)

* join gaps and fixes

* to_parquet with dictionary arrays (#3700)

* added initial support for to_parquet

* updated a comment

* Added the table test, still need to implement tables

* Added support for tables

* moved tests to IO file

* fixed a bug

* Fix gcs issue dict_arrays (#3702)

* added the init

* applied Juan's suggestions

* restore comment

* updated to use Juan's comment

* minor test fixes

* fix MultiIndex scatterv for dict arr

* Fix dict cfuncs in join + support for merging on 1 dict array and 1 string array (#3704)

* fixed typing issues, now there is a segfault

* fixed other python issues

* fixed NA need to fix len 7 now

* fixed 1-indexing

* updated NA handling

* minor refactoring

* fixed for sequential data

* fixed merged impl, need to handle string + dictionary

* support string + dict array join

* fixed the bugs

* fixed AWS test (#3705)

* avoid Pandas error in test_add_series potentially

* skip setitem test

* Fix pivot tests (#3706)

* pivot fixed except n > 1, probably need to fix the shuffle

* fixed pivot except for 1 test

* Fixed the final pivot tests

* fixed gatherv

* cleanup unused code

* Remove more unused code

* avoid issues in test_add_series

* fix removing gatherv in sequential pipeline

* fixed pd.to_datetime for dictionary str arrays (#3708)

* skip unsupported test

* fix scatterv, handle dataframes in decode funcs

* fix decode_if_dict_array optimization

* decode_if_dict_array for SeriesType

* Address comments. And other small tweaks (#3711)

* fix reading dictionary encoded arrays when length is 0

* [WIP] Unify dictionary arrays, need to fix DataFrames (#3710)

* applied ehsan's patch

* cast is supported, need to fix the memory leak

* support for casting arrays

* added the change check

* added DataFrame test

* started writing tests for DataFrame cast

* updated tests for dataframe

* added the null print

* fixed tuple to tuple, but there is a memory leak

* tuple working except memleak, table format needs work

* fix in table, need to debug more

* fixed table format to table format, still a refcount issue

* casts implemented now need to fix memory leaks

* simplified cast handling

* updated test

* Fix NA hash for nullable int arrays (#3714)

* Fix NA hash for nullable int arrays

* minimize changes

* Fix gatherv with read-only arrays (#3715)

* added support for removing gatherv with read only arrays

* fix typo in the test but keep global because it caught the bug

* Fixed the tests/seq version

* removed redundant check

* added null checks

* Made Juan's change

* Apply suggestions from code review

* review comments

* flip flags used for testing

* use dict array in test_unify_dict_string_dataframes

+++ bodo/libs/_array_operations.cpp
@@ -45,12 +45,12 @@ static void array_isin_kernel(array_info* out_arr, array_info* in_arr,
     int64_t len_values = in_values->length;
     std::vector<uint32_t> hashes_values(len_values);
     hash_array(hashes_values.data(), in_values, (size_t)len_values, seed,
-               is_parallel);
+               is_parallel, /*global_dict_needed=*/false);
 
     int64_t len_in_arr = in_arr->length;
     std::vector<uint32_t> hashes_in_arr(len_in_arr);
     hash_array(hashes_in_arr.data(), in_arr, (size_t)len_in_arr, seed,
-               is_parallel);
+               is_parallel, /*global_dict_needed=*/false);
 
     std::function<bool(int64_t, int64_t)> equal_fct =
         [&](int64_t const& pos1, int64_t const& pos2) -> bool {
@@ -506,7 +506,9 @@ table_info* drop_duplicates_table_inner(table_info* in_table, int64_t num_keys,
         key_arrs[iKey] = in_table->columns[iKey];
 
     uint32_t seed = SEED_HASH_CONTAINER;
-    if (hashes == nullptr) hashes = hash_keys(key_arrs, seed, is_parallel);
+    if (hashes == nullptr)
+        hashes = hash_keys(key_arrs, seed, is_parallel,
+                           /*global_dict_needed=*/false);
     HashDropDuplicates hash_fct{hashes};
     KeyEqualDropDuplicates equal_fct{&key_arrs, num_keys};
     UNORD_MAP_CONTAINER<size_t, size_t, HashDropDuplicates,
@@ -660,7 +662,8 @@ table_info* drop_duplicates_keys(table_info* in_table, int64_t num_keys,
  * @return the vector of pointers to be used.
  */
 table_info* drop_duplicates_table(table_info* in_table, bool is_parallel,
-                                  int64_t num_keys, int64_t keep, bool dropna) {
+                                  int64_t num_keys, int64_t keep, bool dropna,
+                                  bool drop_local_first) {
     try {
         // serial case
         if (!is_parallel) {
@@ -670,18 +673,22 @@ table_info* drop_duplicates_table(table_info* in_table, bool is_parallel,
         // parallel case
         // pre reduction of duplicates
         table_info* red_table;
-        red_table = drop_duplicates_table_inner(in_table, num_keys, keep, 2,
-                                                is_parallel, dropna);
+        if (drop_local_first)
+            red_table = drop_duplicates_table_inner(in_table, num_keys, keep, 2,
+                                                    is_parallel, dropna);
+        else
+            red_table = in_table;
         // shuffling of values
         table_info* shuf_table =
             shuffle_table(red_table, num_keys, is_parallel);
         // no need to decref since shuffle_table() steals a reference
-        delete_table(red_table);
+        if (drop_local_first) delete_table(red_table);
         // reduction after shuffling
         // We don't drop NA values again because the first
         // drop_duplicates_table_inner should have already handled this
+        if (drop_local_first) dropna = false;
         table_info* ret_table = drop_duplicates_table_inner(
-            shuf_table, num_keys, keep, 1, is_parallel, false);
+            shuf_table, num_keys, keep, 1, is_parallel, dropna);
         delete_table(shuf_table);
         // returning table
         return ret_table;
